requirements.txt
absl-py==2.1.0
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
fake-useragent==2.0.1
filelock==3.16.1
fonttools==4.55.1
fsspec==2024.10.0
grpcio==1.68.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.4
kiwisolver==1.4.7
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.3
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pip==24.3.1
protobuf==5.29.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
setuptools==75.6.0
six==1.16.0
soupsieve==2.6
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
torch==2.5.1+cu124
torchvision==0.20.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
wheel==0.45.1
zipp==3.21.0
